fix: circular deps check in CI - remove `--summary` flag so exit codes are correct - resolve all circular refs within packages
