+++ packages/cardano-services/src/Program/programs/index.ts
@@ -1,13 +1,4 @@
 export * from './blockfrostWorker';
 export * from './providerServer';
 export * from './txWorker';
-
-/**
- * cardano-services programs
- */
-
-export enum Programs {
-  BlockfrostWorker = 'Blockfrost worker',
-  ProviderServer = 'Provider server',
-  RabbitmqWorker = 'RabbitMQ worker'
-}
+export * from './types';
