+++ packages/cardano-services/src/StakePool/HttpStakePoolMetadata/HttpStakePoolMetadataService.ts
@@ -9,7 +9,8 @@ import { Hash32ByteBase16 } from '@cardano-sdk/crypto';
 import { HexBlob } from '@cardano-sdk/util';
 import { Logger } from 'ts-log';
 import { StakePoolExtMetadataResponse, StakePoolMetadataService } from '../types';
-import { StakePoolMetadataResponse, StakePoolMetadataServiceError, StakePoolMetadataServiceFailure } from '../..';
+import { StakePoolMetadataResponse } from './types';
+import { StakePoolMetadataServiceError, StakePoolMetadataServiceFailure } from './errors';
 import { ValidationError, validate } from 'jsonschema';
 import { getExtMetadataUrl, getSchemaFormat, loadJsonSchema } from './util';
 import { mapToExtendedMetadata } from './mappers';
