+++ packages/dapp-connector/package.json
@@ -37,7 +37,7 @@
     "build:esm": "tsc -p src/tsconfig.json --outDir ./dist/esm --module es2020",
     "build:cjs": "tsc --build src",
     "build": "run-s build:cjs build:esm module-fixup",
-    "circular-deps:check": "madge --circular dist",
+    "circular-deps:check": "madge --circular dist/cjs",
     "module-fixup": "shx cp ../../build/cjs-package.json ./dist/cjs/package.json && cp ../../build/esm-package.json ./dist/esm/package.json",
     "tscNoEmit": "shx echo typescript --noEmit command not implemented yet",
     "cleanup:dist": "shx rm -rf dist",
