+++ packages/util-dev/package.json
@@ -38,7 +38,7 @@
     "build:cjs": "tsc --build src",
     "build:copy": "shx cp -rf src/chainSync/data dist/cjs/chainSync/",
     "build": "run-s build:cjs build:esm module-fixup build:copy",
-    "circular-deps:check": "madge --circular dist",
+    "circular-deps:check": "madge --circular dist/cjs",
     "module-fixup": "shx cp ../../build/cjs-package.json ./dist/cjs/package.json && cp ../../build/esm-package.json ./dist/esm/package.json",
     "tscNoEmit": "shx echo typescript --noEmit command not implemented yet",
     "cleanup:dist": "shx rm -rf dist",
